requirements.txt
asgiref==3.9.1
Django==5.2.4
django-appconf==1.1.0
django-compressor==4.5.1
django-libsass==0.9
libsass==0.23.0
python-dotenv==1.1.1
rcssmin==1.1.2
rjsmin==1.2.2
sqlparse==0.5.3
tzdata==2025.2
